==> 4B_time_to_publication.R <==
#4B_time_to_publication.R

#not run - updated coding saved as .rda
# scoredTrials <- readRDS("data/scoredTrials.rds")
# #prob_adj_threshold <- 0.5 #no longer relevant
# date_of_score_matching <- as.Date.character("2022-11-01")
# possible_pubs = filter(scoredTrials,NCT_num %in% dat_included[['NCT']]) %>% arrange(NCT_num)



source("99_packages.R")
library(survival)
library(ggsurvfit)

# load included studies and scores from matching algorithm
load('data/final_studies_with_meta_R1.rda') 
dat_included = filter(decisions,final_decision=='include') #decision info, including tags
load('data/publication_outcomes_20231204.rda')

# add study start info for all studies
active <- readRDS("data/clintrials_history_active_studies.rds")
complete <- readRDS("data/clintrials_history_completed.rds")
unknown <- readRDS("data/clintrials_history_unknown_stopped.rds")

web_history = bind_rows(active,complete,unknown)
rm(active,complete,unknown)

cal_months = as.character(month(1:12,label=T,abbr=F)) %>% str_c(collapse = '|')
cal_days = paste0('\\b',1:31,'\\b') %>% str_c(collapse = '|')

study_start_dates = web_history %>% unnest(cols=c(field_label,field_value)) %>% filter(field_label %in% c('Study Start','First Submitted thatMet QC Criteria')) %>% spread(field_label,field_value) %>%
  mutate_at('Study Start',~ifelse(. %in% c('Missing','January 2100'),NA,.)) %>% filter(!is.na(`Study Start`),index==1) %>%
  mutate(year = str_extract(string=`Study Start`,pattern='19\\d+|20\\d+'),
         month = str_extract(string=`Study Start`,pattern=cal_months),
         day = str_extract(string=`Study Start`,pattern=cal_days)) %>%
  mutate_at('day',~replace_na(.,'01')) %>% unite("study_start",year:day)

#manual fixes not captured by web history
ad = tibble(NCT=c('NCT01452802','NCT01080157','NCT00435097'),study_start = c('2011_October_01','2010_March_01','2007_February_14'))
study_start_dates = bind_rows(study_start_dates,ad) %>% mutate_at('study_start',~ymd(.))

#add study start to dat_included
dat_included = dat_included %>% left_join(select(study_start_dates,NCT,study_start),by='NCT')
#do the same for posted date
dat_included = dat_included %>% mutate(year = str_extract(string=posted,pattern='19\\d+|20\\d+'),
                                       month = str_extract(string=posted,pattern=cal_months),
                                       day = str_extract(string=posted,pattern=cal_days)) %>%
  mutate_at('day',~replace_na(.,'01')) %>% unite("study_posted",year:day) %>% mutate_at('study_posted',~ymd(.)) %>%
  mutate_at(vars(Development:Validation),~replace_na(.,FALSE)) %>%
  mutate(analysis = case_when(
    (Development==TRUE & Validation==FALSE) ~ 'Development',
    (Development==FALSE & Validation==TRUE) ~ 'Validation',
    (Development==TRUE & Validation==TRUE) ~ 'Development + Validation'),
    outcome = case_when(
      (Prognostic==TRUE & Diagnostic==FALSE) ~ 'Prognostic',
      (Prognostic==FALSE & Diagnostic==TRUE) ~ 'Diagnostic',
      (Prognostic==TRUE & Diagnostic==TRUE) ~ 'Diagnostic + Prognostic'))


dat_times = select(dat_included,NCT:Validation,analysis,outcome,submitted,posted,updated,study_start,study_posted) %>% 
  full_join(select(combined_pmid_results,-c(study_start,study_posted)),by='NCT') #study_start only for ncts with 1+ pubs returned

#exclude duplicate NCT
dat_dup = openxlsx::read.xlsx('data/manual checks/possible duplicates-20230713.xlsx')
exclude_dup = dat_dup %>% filter(duplicate=='y') %>% pull(b) %>% unique()

dat_times = filter(dat_times,!NCT %in% exclude_dup)

#add years to publication
dat_times = dat_times %>% 
  mutate_at(c('pub_date'),~dmy(.)) %>%
  mutate(years_to_publication = as.numeric(pub_date - study_start,'days')/365) %>%
  mutate_at('source',~replace_na(.,'No matches'))

#add outcome for every row; 0= published but no match, 1  = matched pub, 2 = no publications found
dat_times = dat_times %>% mutate(pub_outcome = case_when(match==0 ~ 0,match==1 ~ 1, is.na(match) ~ 2)) 


#publication outcome excluding ML classifier
dat_pubs_1 <- filter(dat_times,pub_outcome==1,source=='NCT linked or investigator submitted') %>% group_by(NCT) %>% 
  slice_min(years_to_publication) %>% ungroup() %>% add_column(flag=1)
res = anti_join(dat_times,dat_pubs_1,by="NCT") %>% distinct(NCT,.keep_all=T) %>% add_column(flag=0)
dat_pubs_1 = bind_rows(dat_pubs_1,res)

#publication outcomes including ML classifier
dat_pubs_2 <- filter(dat_times,pub_outcome==1) %>% group_by(NCT) %>% 
  slice_min(years_to_publication) %>% ungroup() %>% add_column(flag=1)
res = anti_join(dat_times,dat_pubs_2,by="NCT") %>% distinct(NCT,.keep_all=T) %>% add_column(flag=0)
dat_pubs_2 = bind_rows(dat_pubs_2,res)

#pmin(max_time,years_to_publication)
censor_year = 10
last_updated = as.Date('2023-12-04')
#excluding ML
dat_pubs_1 = dat_pubs_1 %>% 
  mutate(
    exposure_time = pmin(censor_year,as.numeric(last_updated-study_start,'days')/365), #take 10 years since study start or up to last_updated, whichever happens first
    status = ifelse(flag==1 & !is.na(years_to_publication) & years_to_publication<exposure_time,1,0),
    event_time_c = case_when(status==1 ~ years_to_publication,status==0 ~ exposure_time)) %>% #because status=0 will still have publications that don't match or were published 10+ years
  mutate_at(vars(Development:Validation),~replace_na(.,FALSE)) %>%
  mutate(analysis = case_when(
    (Development==TRUE & Validation==FALSE) ~ 'Development',
    (Development==FALSE & Validation==TRUE) ~ 'Validation',
    (Development==TRUE & Validation==TRUE) ~ 'Development + Validation'),
    outcome = case_when(
      (Prognostic==TRUE & Diagnostic==FALSE) ~ 'Prognostic',
      (Prognostic==FALSE & Diagnostic==TRUE) ~ 'Diagnostic',
      (Prognostic==TRUE & Diagnostic==TRUE) ~ 'Diagnostic + Prognostic'))


#including ML
dat_pubs_2 = dat_pubs_2 %>% 
  mutate(
    exposure_time = pmin(censor_year,as.numeric(last_updated-study_start,'days')/365), #take 10 years since study start or up to last_updated, whichever happens first
    status = ifelse(flag==1 & !is.na(years_to_publication) & years_to_publication<exposure_time,1,0),
    event_time_c = case_when(status==1 ~ years_to_publication,status==0 ~ exposure_time)) %>% #because status=0 will still have publications that don't match or were published 10+ years
  mutate_at(vars(Development:Validation),~replace_na(.,FALSE)) %>%
  mutate(analysis = case_when(
    (Development==TRUE & Validation==FALSE) ~ 'Development',
    (Development==FALSE & Validation==TRUE) ~ 'Validation',
    (Development==TRUE & Validation==TRUE) ~ 'Development + Validation'),
    outcome = case_when(
      (Prognostic==TRUE & Diagnostic==FALSE) ~ 'Prognostic',
      (Prognostic==FALSE & Diagnostic==TRUE) ~ 'Diagnostic',
      (Prognostic==TRUE & Diagnostic==TRUE) ~ 'Diagnostic + Prognostic'))


save(dat_included,dat_times,dat_pubs_1,dat_pubs_2,file='data/publications/times_to_publication_20231204.rda')

#cumulative incidence functions 

#without ML classifier
cuminc_fit =with(dat_pubs_1,cuminc(ftime=event_time_c,fstatus=status,cencode=0))
gdata_1 = ggcompetingrisks(cuminc_fit,conf.int = T)$data %>% mutate_at('name',~str_remove_all(.,' 1')) %>% distinct()

g1 = ggplot(gdata_1,aes(x=time,y=est,ymin=est-1.96*std,ymax=est+1.96*std))+geom_path()+geom_ribbon(alpha=.2)+
  expand_limits(y=0.4)+
  scale_x_continuous('Years since study start',breaks=seq(0,10,1),labels=0:10,expand = c(0, 0), limits = c(0, NA))+
  scale_y_continuous('Cumulative incidence,\ntime to first matching publication',breaks=seq(0,0.5,0.05),expand = c(0, 0), limits = c(0, NA))+theme_bw()+
  theme(legend.background = element_blank(),text=element_text(size=12),axis.text=element_text(size=12),panel.grid.minor = element_blank(),plot.margin = margin(1,1,1,1, "cm"))

#with ML classifier
cuminc_fit =with(dat_pubs_2,cuminc(ftime=event_time_c,fstatus=status,cencode=0))
gdata_2 = ggcompetingrisks(cuminc_fit,conf.int = T)$data %>% mutate_at('name',~str_remove_all(.,' 1')) %>% distinct()

g2 = ggplot(gdata_2,aes(x=time,y=est,ymin=est-1.96*std,ymax=est+1.96*std))+geom_path()+geom_ribbon(alpha=.2)+
  expand_limits(y=0.4)+
  scale_x_continuous('Years since study start',breaks=seq(0,10,1),labels=0:10,expand = c(0, 0), limits = c(0, NA))+
  scale_y_continuous('Cumulative incidence,\ntime to first matching publication',breaks=seq(0,0.5,0.05),expand = c(0, 0), limits = c(0, NA))+theme_bw()+
  theme(legend.background = element_blank(),text=element_text(size=12),axis.text=element_text(size=12),panel.grid.minor = element_blank(),plot.margin = margin(1,1,1,1, "cm"))


#stratify all matches by study start year
pubs_by_era = dat_pubs_2 %>% mutate(strat = case_when(
  between(year(study_start),1990,2010)~'pre-2010',
  between(year(study_start),2011,2015)~'2011 to 2015',
  between(year(study_start),2016,2019)~'2016 to 2019',
  TRUE ~ '2020 to 2022'
)) %>% count(strat) %>% mutate(strat_label = paste0(strat,' (n = ',n,')')) %>% arrange(n)

dat_pubs_2 = dat_pubs_2 %>% mutate(strat = case_when(between(year(study_start),1990,2010)~'pre-2010',
                                                     between(year(study_start),2011,2015)~'2011 to 2015',
                                                     between(year(study_start),2016,2019) ~ '2016 to 2019',
                                                     TRUE ~ '2020 to 2022')) %>% left_join(pubs_by_era,by='strat')



cuminc_fit_strat =with(dat_pubs_2,cuminc(ftime=event_time_c,fstatus=status,cencode=0,group=strat_label))
gdata_strat = ggcompetingrisks(cuminc_fit_strat,conf.int = T)$data %>% mutate_at('name',~str_remove_all(.,' 1$')) %>% distinct() %>%
  mutate_at('name',~factor(.,levels=c('pre-2010 (n = 68)','2011 to 2015 (n = 128)','2016 to 2019 (n = 324)','2020 to 2022 (n = 408)')))

g3 = ggplot(gdata_strat,aes(x=time,y=est,group=name,colour=name,fill=name))+geom_path(linewidth=1)+
  expand_limits(y=0.4)+
  scale_x_continuous('Years since study start',breaks=seq(0,10,1),labels=0:10,expand = c(0, 0), limits = c(0, NA))+
  scale_y_continuous('Cumulative incidence,\ntime to first relevant publication',breaks=seq(0,0.6,0.05),expand = c(0, 0), limits = c(0, NA))+theme_bw()+
  theme(legend.background = element_blank(),legend.position = c(0.25,0.75),legend.title=element_blank(),legend.text=element_text(size=12),text=element_text(size=12),axis.text=element_text(size=12),panel.grid.minor = element_blank(),
        plot.margin = margin(1,1,1,1, "cm"))


png('manuscript/figures/Figure4.png',width=8,height=8,units='in',res=300)
ggarrange(g2,g3,nrow=2,ncol=1,labels=LETTERS[1:2])
dev.off()

#censoring figure for the supplement
cuminc_fit_cens =with(dat_pubs_2,cuminc(ftime=event_time_c,fstatus=status,cencode=1))
gdata_cens = ggcompetingrisks(cuminc_fit_cens,conf.int = T)$data %>% mutate_at('name',~str_remove_all(.,' 0')) %>% distinct()

g4 = ggplot(gdata_cens,aes(x=time,y=est,ymin=est-1.96*std,ymax=est+1.96*std))+geom_path()+geom_ribbon(alpha=.2)+
  expand_limits(y=1)+
  scale_x_continuous('Years since study start',breaks=seq(0,120,12),labels=0:10,expand = c(0, 0), limits = c(0, NA))+
  scale_y_continuous('Cumulative incidence of censoring',breaks=seq(0,1,0.1),expand = c(0, 0), limits = c(0, NA))+theme_bw()+
  theme(legend.background = element_blank(),text=element_text(size=12),axis.text=element_text(size=12),panel.grid.minor = element_blank(),plot.margin = margin(1,1,1,1, "cm"))


png('manuscript/supplement/FigureS7.png',width=8,height=8,units='in',res=300)
g4
dev.off()

##
gdata = bind_rows(list('NCT-linked + investigator-submitted'=gdata_1,'NCT-linked + investigator-submitted + ML classifier'=gdata_2),.id='group')


gall = ggplot(gdata,aes(x=time,y=est,ymin=est-1.96*std,ymax=est+1.96*std,group=group,colour=group,fill=group))+geom_path()+geom_ribbon(alpha=.2)+
  expand_limits(y=0.4)+
  scale_x_continuous('Years since study start',breaks=seq(0,10,1),labels=0:10,expand = c(0, 0), limits = c(0, NA))+
  scale_y_continuous('Cumulative incidence,\ntime to first matching publication',breaks=seq(0,0.4,0.05),expand = c(0, 0), limits = c(0, NA))+theme_bw()+
  theme(legend.background = element_blank(),legend.position = c(0.35,0.85),legend.title=element_blank(),legend.text=element_text(size=12),text=element_text(size=12),axis.text=element_text(size=12),panel.grid.minor = element_blank(),plot.margin = margin(1,1.5,1,1, "cm"))
png('manuscript/supplement/FigureS3.png',width=10,height=8,units='in',res=300)
gall
dev.off()


#
#
# #pre2020
# cuminc_fit =with(filter(dat_pubs_1,year(study_start)<2020),cuminc(ftime=event_time_c,fstatus=status_1,cencode=0))
# gdata_1 = ggcompetingrisks(cuminc_fit,conf.int = T)$data %>% mutate_at('name',~str_remove_all(.,' 1')) %>% distinct()
# cuminc_fit =with(filter(dat_pubs_1,year(study_start)<2020),cuminc(ftime=event_time_c,fstatus=status_2,cencode=0))
# gdata_2 = ggcompetingrisks(cuminc_fit,conf.int = T)$data %>% mutate_at('name',~str_remove_all(.,' 1')) %>% distinct()
# gdata = bind_rows(list('NCT-linked + investigator-submitted'=gdata_1,'NCT-linked + investigator-submitted + ML classifier'=gdata_2),.id='group')
#
# g2020 = ggplot(gdata,aes(x=time,y=est,ymin=est-1.96*std,ymax=est+1.96*std,group=group,colour=group,fill=group))+geom_path()+geom_ribbon(alpha=.2)+
#   expand_limits(y=0.3)+
#   scale_x_continuous('Years since study start',breaks=seq(0,120,12),labels=0:10,expand = c(0, 0), limits = c(0, NA))+
#   scale_y_continuous('Cumulative incidence\ntime to first relevant publication',breaks=seq(0,0.3,0.05),expand = c(0, 0), limits = c(0, NA))+theme_bw()+
#   theme(legend.background = element_blank(),legend.title=element_blank(),legend.text=element_text(size=12),text=element_text(size=12),axis.text=element_text(size=12),panel.grid.minor = element_blank(),
#         plot.margin = margin(1,1.5,1,1, "cm"))
#
#
# #pre-2017
# cuminc_fit =with(filter(dat_pubs_1,year(study_start)<2017),cuminc(ftime=event_time_c,fstatus=status_1,cencode=0))
# gdata_1 = ggcompetingrisks(cuminc_fit,conf.int = T)$data %>% mutate_at('name',~str_remove_all(.,' 1')) %>% distinct()
# cuminc_fit =with(filter(dat_pubs_1,year(study_start)<2017),cuminc(ftime=event_time_c,fstatus=status_2,cencode=0))
# gdata_2 = ggcompetingrisks(cuminc_fit,conf.int = T)$data %>% mutate_at('name',~str_remove_all(.,' 1')) %>% distinct()
# gdata = bind_rows(list('NCT-linked + investigator-submitted'=gdata_1,'NCT-linked + investigator-submitted + ML classifier'=gdata_2),.id='group')
#
#
# g2017 = ggplot(gdata,aes(x=time,y=est,ymin=est-1.96*std,ymax=est+1.96*std,group=group,colour=group,fill=group))+geom_path()+geom_ribbon(alpha=.2)+
#   expand_limits(y=0.3)+
#   scale_x_continuous('Years since study start',breaks=seq(0,120,12),labels=0:10,expand = c(0, 0), limits = c(0, NA))+
#   scale_y_continuous('Cumulative incidence\ntime to first relevant publication',breaks=seq(0,0.3,0.05),expand = c(0, 0), limits = c(0, NA))+theme_bw()+
#   theme(legend.background = element_blank(),legend.title=element_blank(),legend.text=element_text(size=12),text=element_text(size=12),axis.text=element_text(size=12),panel.grid.minor = element_blank(),
#         plot.margin = margin(1,1.5,1,1, "cm"))
#

# png('manuscript/supplement/FigureS3.png',width=8,height=12,units='in',res=300)
# ggarrange(gall,g2017,g2020,nrow=3,ncol=1,labels=c('All records','2000 to 2017','2000 to 2020'),common.legend = T,legend='bottom')
# dev.off()



#study type

cuminc_fit_d =with(filter(dat_pubs_2,Development==TRUE),cuminc(ftime=event_time_c,fstatus=status,cencode=0))
cuminc_fit_dv =with(filter(dat_pubs_2,Development==TRUE,Validation==TRUE),cuminc(ftime=event_time_c,fstatus=status,cencode=0))
cuminc_fit_v =with(filter(dat_pubs_2,Development==FALSE,Validation==TRUE),cuminc(ftime=event_time_c,fstatus=status,cencode=0))


gdata_d = ggcompetingrisks(cuminc_fit_d,conf.int = T)$data %>% mutate_at('name',~str_remove_all(.,' 1')) %>% distinct()
gdata_dv = ggcompetingrisks(cuminc_fit_dv,conf.int = T)$data %>% mutate_at('name',~str_remove_all(.,' 1')) %>% distinct()
gdata_v = ggcompetingrisks(cuminc_fit_v,conf.int = T)$data %>% mutate_at('name',~str_remove_all(.,' 1')) %>% distinct()

gdata = bind_rows(list('Development only'=gdata_d,'Development + Validation'=gdata_dv,'Validation only'=gdata_v),.id='group')

png('manuscript/supplement/FigureS4.png',width=8,height=8,units='in',res=300)
ggplot(gdata,aes(x=time,y=est,ymin=est-1.96*std,ymax=est+1.96*std,group=group,colour=group,fill=group))+geom_path()+geom_ribbon(alpha=.2)+
  expand_limits(y=0.55)+
  scale_x_continuous('Years since study start',breaks=seq(0,10,1),labels=0:10,expand = c(0, 0), limits = c(0, NA))+
  scale_y_continuous('Cumulative incidence,\ntime to first relevant publication',breaks=seq(0,0.55,0.05),expand = c(0, 0), limits = c(0, NA))+theme_bw()+
  theme(legend.background = element_blank(),legend.position = c(0.2,0.85),legend.title=element_blank(),legend.text=element_text(size=12),text=element_text(size=12),axis.text=element_text(size=12),panel.grid.minor = element_blank(),
        plot.margin = margin(1,1.5,1,1, "cm"))
dev.off()


#model outcome

cuminc_fit_d =with(filter(dat_pubs_2,Diagnostic==TRUE),cuminc(ftime=event_time_c,fstatus=status,cencode=0))
cuminc_fit_p =with(filter(dat_pubs_2,Prognostic==TRUE),cuminc(ftime=event_time_c,fstatus=status,cencode=0))


gdata_d = ggcompetingrisks(cuminc_fit_d,conf.int = T)$data %>% mutate_at('name',~str_remove_all(.,' 1')) %>% distinct()
gdata_p = ggcompetingrisks(cuminc_fit_p,conf.int = T)$data %>% mutate_at('name',~str_remove_all(.,' 1')) %>% distinct()

gdata = bind_rows(list('Diangostic'=gdata_d,'Prognostic'=gdata_p),.id='group')

png('manuscript/supplement/FigureS5.png',width=8,height=8,units='in',res=300)
ggplot(gdata,aes(x=time,y=est,ymin=est-1.96*std,ymax=est+1.96*std,group=group,colour=group,fill=group))+geom_path()+geom_ribbon(alpha=.2)+
  expand_limits(y=0.45)+
  scale_x_continuous('Years since study start',breaks=seq(0,10,1),labels=0:10,expand = c(0, 0), limits = c(0, NA))+
  scale_y_continuous('Cumulative incidence,\ntime to first relevant publication',breaks=seq(0,0.45,0.05),expand = c(0, 0), limits = c(0, NA))+theme_bw()+
  theme(legend.background = element_blank(),legend.position = c(0.2,0.85),legend.title=element_blank(),legend.text=element_text(size=12),text=element_text(size=12),axis.text=element_text(size=12),panel.grid.minor = element_blank(),
        plot.margin = margin(1,1.5,1,1, "cm"))
dev.off()








#########################################
#not run  
# dat_times_c = dat_times %>% select(NCT,overall_status,study_start,PMID,NCT.Link,Trial.Results,prob_adj) %>% mutate_at('PMID',~as.character(.))
# save(dat_times_c,file='data/scored_trials_with_meta.rda')
# 
# load('data/pubmed_publication_data.rda')

#MEDLINE search terms form Geersing et al (2021) - modified for pubmed (recheck later)
##inui original filter
# (Validat$ OR Predict$.ti. OR Rule$) OR (Predict$ AND (Outcome$ OR Risk$ OR Model$)) OR 
#((History OR Variable$ OR Criteria OR Scor$ OR Characteristic$ OR Finding$ OR Factor$) AND (Predict$ OR Model$ OR Decision$ OR Identif$ OR Prognos$)) OR (Decision$ AND (Model$ OR Clinical$ OR Logistic Models/))
# OR (Prognostic AND (History OR Variable$ OR Criteria OR Scor$ OR Characteristic$ OR Finding$ OR Factor$ OR Model$))
# 
# 
# mod_ingui_filter <- c('validat(.*)','predict(.*)','rule','predict(.*)outcome','predict(.*)model','predict(.*)outcome','predict(.*)risk','develop(.*)')
# mod_update_filter <- c("stratification","roc curve","discriminat(.*)","c(.*)statistic","area under the curve","auc","calibration","indices","algorithm","multivariable")
# 
# mod_ingui_filter = str_c(paste0('\\b',mod_ingui_filter,'\\b'),collapse = '|')
# mod_update_filter = str_c(paste0('\\b',mod_update_filter,'\\b'),collapse = '|')
# search_str = paste(mod_ingui_filter,mod_update_filter,sep='|')
# 
# 
# cited_pubs = dat_times_c %>% inner_join(pubmed_tiab,by=c('PMID')) %>% filter(NCT.Link==1|Trial.Results==1,!is.na(title))
# 
# notcited_pubs = filter(pubmed_tiab,!NCT_num %in% cited_pubs$NCT) %>% filter(grepl(search_str,abstract,ignore.case=T)|grepl(search_str,title,ignore.case=T)) %>% distinct()
# notcited_pubs = inner_join(dat_times_c,notcited_pubs,by='PMID') %>% filter(prob_adj>0.05)
# 
# save(cited_pubs,notcited_pubs,search_str,file='data/manual checks/pubmed matches v1.rda')
# write.table(select(cited_pubs,NCT,PMID,title),file='data/manual checks/pubmed matches nct clin v1.csv',sep=',',row.names=F)
# write.table(select(notcited_pubs,NCT,PMID,title,prob_adj),file='data/manual checks/pubmed matched algorithm v1.csv',sep=',',row.names=F)


#cited_pubs %>% mutate(time_to_pub = as.numeric(as.Date(pub_date) - study_start,'days')*(12/365)) %>% summarise(median(time_to_pub,na.rm=T))
